refactor(secretman): split crypt workflows into focused files
Move encryption, decryption, refresh, and env diff logic out of crypt.go
into dedicated files.

Extract the per-file encryption flow into helper functions to make the
main encryption command easier to read while preserving the existing
SOPS behavior.
